+++ src/constants.ts
@@ -7,5 +7,3 @@ export const TITLE_SHOW_LOG = "Show Log";
 
 export const FS_EVENT_TYPE_REMOVE = "rename";
 export const FS_EVENT_TYPE_CHANGE = "change";
-
-export const TIME_CACHE_LIFETIME = 240_000;